try fix error in macos
