+++ base-coroutine/src/lib.rs
@@ -24,9 +24,6 @@ pub use work_steal::*;
 mod random;
 
 #[allow(dead_code)]
-mod work_steal_v2;
-
-#[allow(dead_code)]
 pub mod scheduler;
 
 pub use scheduler::*;
